Revert "Proof of concept outline for VE-VE exchanges&joins"
This reverts commit 2e7bff59
