+++ .github/workflows/benchmark-collector.yml
@@ -1,10 +1,11 @@
 ---
 name: "TPC-H Central runner"
 env:
-  INPUT_exchange-strategy: "${{ github.event.inputs.exchange-strategy }}"
   INPUT_extra: "${{ github.event.inputs.extra }}"
   INPUT_fail-fast: "${{ github.event.inputs.fail-fast }}"
-  INPUT_join-strategy: "${{ github.event.inputs.join-strategy }}"
+  INPUT_filter-on-ve: "${{ github.event.inputs.filter-on-ve }}"
+  INPUT_join-on-ve: "${{ github.event.inputs.join-on-ve }}"
+  INPUT_pass-through-project: "${{ github.event.inputs.pass-through-project }}"
   INPUT_query: "${{ github.event.inputs.query }}"
   INPUT_scale: "${{ github.event.inputs.scale }}"
   INPUT_serializer: "${{ github.event.inputs.serializer }}"
@@ -13,14 +14,6 @@ env:
 "on":
   workflow_dispatch:
     inputs:
-      exchange-strategy:
-        type: "choice"
-        description: "Exchange strategy"
-        options:
-        - "spark"
-        - "vector-host"
-        - "none"
-        default: "spark"
       extra:
         type: "input"
         description: "Extra command line arguments to add to Spark"
@@ -28,14 +21,18 @@ env:
         type: "boolean"
         description: "Fail Fast"
         default: "true"
-      join-strategy:
-        type: "choice"
-        description: "Join strategy"
-        options:
-        - "-"
-        - "host"
-        - "spark"
-        default: "-"
+      filter-on-ve:
+        type: "boolean"
+        description: "Filter on VE"
+        default: "true"
+      join-on-ve:
+        type: "boolean"
+        description: "Join on VE"
+        default: "false"
+      pass-through-project:
+        type: "boolean"
+        description: "Pass-through in projection"
+        default: "false"
       query:
         type: "choice"
         description: "Query to use"
