+++ src/main/scala/com/nec/spark/planning/aggregation/VeHashExchangePlan.scala
@@ -13,17 +13,16 @@ import com.nec.spark.SparkCycloneExecutorPlugin.metrics.{
   measureRunningTime,
   registerFunctionCallTime
 }
-import com.nec.ve.exchange.VeExchangeStrategy
+import com.nec.ve.VeRDD.RichKeyedRDDL
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.execution.{SparkPlan, UnaryExecNode}
 
-case class VeHashExchangePlan(
-  exchangeFunction: VeFunction,
-  child: SparkPlan,
-  veExchangeStrategy: VeExchangeStrategy
-) extends UnaryExecNode
+import java.time.{Duration, Instant}
+
+case class VeHashExchangePlan(exchangeFunction: VeFunction, child: SparkPlan)
+  extends UnaryExecNode
   with SupportsVeColBatch
   with LazyLogging
   with PlanCallsVeFunction
@@ -31,7 +30,6 @@ case class VeHashExchangePlan(
 
   import com.nec.spark.SparkCycloneExecutorPlugin.veProcess
   import OriginalCallingContext.Automatic._
-  import veExchangeStrategy.Implicits._
 
   override def executeVeColumnar(): RDD[VeColBatch] = child
     .asInstanceOf[SupportsVeColBatch]
