+++ tpcbench-run/src/main/scala/sc/RunOptions.scala
@@ -33,8 +33,8 @@ final case class RunOptions(
   filterOnVe: Boolean,
   passThroughProject: Boolean,
   failFast: Boolean,
-  joinStrategy: Option[String],
-  exchangeStrategy: Option[String]
+  joinOnVe: Boolean,
+  exchangeOnVe: Boolean
 ) {
 
   def enhanceWithEnv(env: Map[String, String]): RunOptions = env
@@ -65,7 +65,9 @@ final case class RunOptions(
       ("spark.com.nec.spark.pass-through-project", passThroughProject),
       ("spark.com.nec.spark.project-on-ve", projectOnVe),
       ("spark.com.nec.spark.filter-on-ve", filterOnVe),
-      ("spark.com.nec.spark.fail-fast", failFast)
+      ("spark.com.nec.spark.exchange-on-ve", exchangeOnVe),
+      ("spark.com.nec.spark.fail-fast", failFast),
+      ("spark.com.nec.spark.join-on-ve", joinOnVe)
     ).map { case (k, v) => (k, v.toString) }
   }
 
@@ -88,12 +90,9 @@ final case class RunOptions(
       case ("ve-log-debug", v)                             => copy(veLogDebug = v == "on" || v == "true")
       case ("pass-through-project", v)                     => copy(passThroughProject = v == "on" || v == "true")
       case ("fail-fast", v)                                => copy(failFast = v == "on" || v == "true")
-      case ("join-strategy", v) =>
-        copy(joinStrategy = Option(v).filter(_ != "-").filter(_.nonEmpty))
-      case ("exchange-strategy", v) =>
-        copy(exchangeStrategy = Option(v).filter(_ != "-").filter(_.nonEmpty))
-      case ("filter-on-ve", v)         => copy(filterOnVe = v == "on" || v == "true")
-      case ("kernel-directory", newkd) => copy(kernelDirectory = Some(newkd))
+      case ("join-on-ve", v)                               => copy(joinOnVe = v == "on" || v == "true")
+      case ("filter-on-ve", v)                             => copy(filterOnVe = v == "on" || v == "true")
+      case ("kernel-directory", newkd)                     => copy(kernelDirectory = Some(newkd))
     }
   }
 
@@ -155,14 +154,6 @@ final case class RunOptions(
       "--conf",
       "spark.executorEnv.VE_OMP_NUM_THREADS=1"
     ) ++ {
-      joinStrategy.toList.flatMap { str =>
-        List("--conf", s"spark.com.nec.spark.join-strategy=${str}")
-      }
-    } ++ {
-      exchangeStrategy.toList.flatMap { str =>
-        List("--conf", s"spark.com.nec.spark.exchange-strategy=${str}")
-      }
-    } ++ {
       if (codeDebug)
         List("--conf", "spark.com.nec.spark.ncc.debug=true")
       else Nil
@@ -233,12 +224,12 @@ object RunOptions {
     enableVeSorting = true,
     projectOnVe = true,
     filterOnVe = true,
-    exchangeStrategy = None,
+    exchangeOnVe = true,
     codeDebug = false,
     useCyclone = true,
     passThroughProject = false,
     failFast = true,
-    joinStrategy = None
+    joinOnVe = false
   )
 
   import scala.collection.JavaConverters._
